requirements.txt
# SPDX-License-Identifier: AGPL-3.0-or-later
# This file is part of MLX DeepSeek-OCR and is covered by AGPL-3.0.
# See the LICENSE file in the project root for full license text:
# https://www.gnu.org/licenses/agpl-3.0.en.html

Flask==3.0.0
mlx-vlm==0.3.5
mlx>=0.20.0
Pillow>=10.3.0
Werkzeug==3.0.1
PyMuPDF
opencv-python>=4.10.0
